Changed amount column to float in transactions table

+++ db/migrate/20230217122707_create_transactions.rb
@@ -1,7 +1,7 @@
 class CreateTransactions < ActiveRecord::Migration[7.0]
   def change
     create_table :transactions do |t|
-      t.decimal :amount, null: false, default: 0.0
+      t.float :amount, null: false, default: 0.0
       t.integer :status, null: false, default: 0
       t.integer :transaction_type, null: false
       t.string :transaction_reference, null: false, unique: true, index: true
